+++ types/artist.ts
@@ -1,25 +1,5 @@
 import type { Image } from "./image";
 
-export type TopArtistsAttributes = {
-	user: string;
-	totalPages: string;
-	page: string;
-	total: string;
-	perPage: string;
-};
-
-export type ArtistInfoStats = {
-	listeners: string;
-	playcount: string;
-	userplaycount: string;
-};
-
-export type ArtistInfoLink = {
-	"#text": string;
-	rel: string;
-	href: string;
-};
-
 export type ArtistAttributes = {
 	rank: string;
 };
