+++ types/index.ts
@@ -1,7 +1,5 @@
 export * from "./api";
 export * from "./artist";
 export * from "./image";
-export * from "./tag";
 export * from "./theme";
 export * from "./track";
-export * from "./widgets";
