--- types/tag.ts
@@ -1,4 +1,0 @@
-export type Tag = {
-	name: string;
-	url: string;
-};
