--- types/widgets/lastfm.ts
@@ -1,15 +1,0 @@
-export type LastFmWidgetArtist = {
-	name: string;
-	playcount: string;
-};
-
-export type LastFmWidgetNowPlaying = {
-	trackName: string;
-	artistName: string;
-	albumTitle: string;
-	encodedTrackImage: string;
-};
-
-export type LastFmWidgetData = LastFmWidgetNowPlaying & {
-	topWeeklyArtists: LastFmWidgetArtist[];
-};
